+++ neuron.go
@@ -6,16 +6,16 @@ import (
 	// "github.com/proxypoke/vector"
 )
 
-type activationFunction func(float32) float32
+type activationFunction func(float64) float64
 
 type Neuron struct {
-	Bias               float32
+	Bias               float64
 	ActivationFunction activationFunction
 	Node
 }
 
 	
-func (neuron *Neuron) computeOutput(weightedInputs []*weightedInput) float32 {
+func (neuron *Neuron) computeOutput(weightedInputs []*weightedInput) float64 {
 
     /*
     reduce_function = fn({inputs, weights}, acc) ->
